refactor: move slate plugins from src/plugins/text to src/components/tiles/text/plugins
Per reviewer feedback: src/plugins/ is for CLUE tile plugins (dynamically
loaded per unit), not Slate editor plugins (always loaded with the text
tile). Moves highlights-plugin, link-plugin, and highlight-registry-context
to src/components/tiles/text/plugins/ where they belong.

Updates all import paths across 10 files.
