+++ src/components/tiles/text/plugins/highlight-registry-context.ts
